SupbrocessServiceManager Logging
We had an issue with logs filling up the disk and causing a firestorm of
problems resulting in an absurd number of SAPI pulls.
SubprocessServiceManager now has a log quota and clean-up logic along
with using RotatingFileHandler for logging to files which guarantees:
- log files will be at most X bytes (default 100MB) with excess logs
  being moved to logile.1, logfile.2, ..., .k. This allows us to
  prune logs even from an active service by deleting some of its backup
  logs (the 'head' log of an active service may never be deleted)
- a predetermined maxium number of .k backups is allowed, guaranteeing
  an upperbound on an active process' log size

The cleanup logic is as follows:
- delete the fewest log files to get within our log quota (default 20G)
- start by deleting from services that have logs for at least two
  terminated instances, starting from the oldest logfile
- continue by deleting from services that have logs for at least ONE
  terminated instance, starting from the oldest logfile
- continue by deleting logfiles from the active instances, starting
  from the service whose logs are the largest.

The RotatingFileHandler is also used for the odb_server logs, instead of
logging to the terminal and relying on the caller to forward stdout (and
stderr). We don't have an instanceID to use for these logs as we do for
service logs, neither can we use the PID (via os.getpid()) because when
run through docker the service_manager.py process gets one of two PIDs
depending on boot timing, which resulted in logs getting overwitten when
moved to old/ or appended when getting the same PID as the last
instance. So instead we use the integer part of the timestamp when the
service_manager.py process is started.

+++ object_database/service_manager/ServiceManager.py
@@ -183,26 +183,36 @@ class ServiceManager(object):
 
     @staticmethod
     def startService(serviceName, targetCount=1):
-        service = service_schema.Service.lookupOne(name=serviceName)
+        service = service_schema.Service.lookupUnique(name=serviceName)
+        if service is None:
+            raise TypeError(f"Unknown service '{serviceName}'")
+
         service.target_count = targetCount
 
     @staticmethod
-    def waitRunning(db, serviceName, timeout=5.0) -> bool:
+    def waitRunning(db, serviceName, timeout=5.0, targetCount=1, exact=False) -> bool:
         def isRunning():
-            service = service_schema.Service.lookupAny(name=serviceName)
+            service = service_schema.Service.lookupUnique(name=serviceName)
             if not service:
                 return False
 
             instances = service_schema.ServiceInstance.lookupAll(service=service)
-            if any(inst.isRunning() for inst in instances):
+            if exact and sum(1 for inst in instances if inst.isRunning()) == targetCount:
                 return True
-            return False
+
+            elif not exact and any(inst.isRunning() for inst in instances):
+                return True
+
+            else:
+                return False
 
         return db.waitForCondition(isRunning, timeout)
 
     @staticmethod
     def stopService(serviceName):
-        service = service_schema.Service.lookupOne(name=serviceName)
+        service = service_schema.Service.lookupUnique(name=serviceName)
+        if service is None:
+            raise TypeError(f"Unknown service '{serviceName}'")
         service.target_count = 0
 
     @staticmethod
@@ -302,11 +312,15 @@ class ServiceManager(object):
                 if not instance.service.exists():
                     orphans.append(instance)
 
-        self.stopServices(orphans)
-        with self.db.transaction():
-            for instance in orphans:
-                if instance.exists():
-                    instance.delete()
+        if len(orphans) > 0:
+            self._logger.warning(
+                f"Removing {len(orphans)} orphaned ServiceInstance instances."
+            )
+            self.stopServices(orphans)
+            with self.db.transaction():
+                for instance in orphans:
+                    if instance.exists():
+                        instance.delete()
 
     @revisionConflictRetry
     def collectDeadHosts(self):
@@ -316,9 +330,14 @@ class ServiceManager(object):
                 instances = service_schema.ServiceInstance.lookupAll(host=serviceHost)
 
                 if not serviceHost.connection.exists():
+                    self._logger.warning(
+                        f"Collecting dead host '{serviceHost.hostname}' "
+                        f"and its {len(instances)} instances."
+                    )
                     for sInst in instances:
                         sInst.delete()
                     serviceHost.delete()
+
                 else:
                     actualRam = sum([i.service.gbRamUsed for i in instances if i.isActive()])
                     actualCores = sum([i.service.coresUsed for i in instances if i.isActive()])
@@ -352,6 +371,10 @@ class ServiceManager(object):
                             serviceInstance.service.lastFailureReason = (
                                 serviceInstance.failureReason
                             )
+                    self._logger.warning(
+                        f"Removing service instance with state='{serviceInstance.state}' "
+                        f"and failureReason='{serviceInstance.failureReason}'."
+                    )
                     serviceInstance.delete()
 
     def redeployServicesIfNecessary(self):
@@ -458,6 +481,7 @@ class ServiceManager(object):
             instance = service_schema.ServiceInstance(
                 service=service, host=host, state="Booting", start_timestamp=time.time()
             )
+            self._logger.info(f"Added Booting instance for service '{service.name}'.")
 
             actual_records.append(instance)
 
@@ -467,6 +491,7 @@ class ServiceManager(object):
         while service.effectiveTargetCount() < len(actual_records):
             sInst = actual_records.pop()
             sInst.triggerShutdown()
+            self._logger.info(f"Triggered shutdown for service '{sInst.service.name}'.")
 
     def instanceRecordsToBoot(self):
         res = []
